feat: Tell user they need to pip install clang.

+++ barretenberg/scripts/bindgen.sh
@@ -1,6 +1,15 @@
 #!/usr/bin/env bash
 set -eu
 
+if ! pip list | grep -E 'clang\s+16.0.6' > /dev/null; then
+  echo "You need to install python clang: pip install clang==16.0.6"
+  exit 1
+fi
+
 #find ./cpp/src -type f -name "c_bind*.hpp" | ./scripts/decls_json.py > exports.json
 cat ./scripts/c_bind_files.txt | ./scripts/decls_json.py > exports.json
-(cd ./ts && yarn node --loader ts-node/esm ./src/bindgen/index.ts ../exports.json > ./src/barretenberg_api/index.ts)
+(
+  cd ./ts && \
+  yarn node --loader ts-node/esm ./src/bindgen/index.ts ../exports.json > ./src/barretenberg_api/index.ts && \
+  yarn prettier -w ./src/barretenberg_api/index.ts
+)
